Rename a scripts (#773)
