+++ .github/workflows/performance-test.yaml
@@ -39,12 +39,12 @@ on:
             testSuitePath:
                 description: 'Path to test suite to run'
                 required: true
-                default: 'test/k6/src/tests/orders-org-no.js'
+                default: 'test/k6/src/tests/orders_org_no.js'
                 type: choice
                 options:
-                - 'test/k6/src/tests/orders-org-no.js'
-                - 'test/k6/src/tests/orders-email.js'
-                - 'test/k6/src/tests/orders-sms.js'
+                - 'test/k6/src/tests/orders_org_no.js'
+                - 'test/k6/src/tests/orders_email.js'
+                - 'test/k6/src/tests/orders_sms.js'
 
 permissions:
   id-token: write
